+++ agenciaturismo/src/main/java/br/senac/tads3/pi03b/gruposete/servlets/CadastroHotelServlet.java
@@ -53,7 +53,7 @@ public class CadastroHotelServlet extends HttpServlet {
 //            erro = true;
 //            request.setAttribute("erroQuantidade_hospedes", true);
 //        }
-        double preco = Double.parseDouble(request.getParameter("preco"));
+        float preco = Float.parseFloat(request.getParameter("preco"));
 //        if (preco < 0) {
 //            erro = true;
 //            request.setAttribute("erroPreco", true);
